Revamp org-roam-template processing.
Org-roam templates are used for two purposes:

- when calling org-roam-capture, and
- when visiting a node that does not exist.

This means that the name of the destination is always being queried, even if it not used
in the template.

This patch does several things:

1. When doing org-roam-capture, it asks for the destination  only
   if the template requests it.

2. If visiting a node that does not exist, a node is created and passed to the templates.

Perhaps more importantly, the code has been refactored and the processing of the templates has
less cloned code. It hopefully makes it easier to maintain.

Some things to note:

- org-roam-capture--node was incorrectly described as only not-nil during a template creation.
  That was not true. This variable is not re-entrant.

- when visiting a node that does not exist, the user is asked for a template. The template
  might never use this node (e.g. if the destination is node)