Add create-file option to templates.
Templates should be of two types:

1. those that create files, and
2. those that do not create files.

This option allows to check the intent of the template.

It has two values:

yes -> only allow the template to create a new file
no  -> template will only operate on files that exist, it will not create it

if the option is not available, the template can be used for both.

Note that this option only applies to "file" type templates.
All node templates require the node to exist.