Add support for functions in the specification of a path.
A path can be a string or a function. If it is a function, it is called
to create the path. For example, the following template will
call the lambda when this template is called. The lambda will
ask for a directory name and prepend it to the slug.

(add-to-list 'org-roam-capture-templates
             `("L" "course lecture" plain
               "%?"
               :target (file+head (lambda ()
                               (concat (dmg-directory-ask) "/${slug}.org"))
                             ,(eval (dmg-roam-template "lecture.txt")))
               :create-file yes
              :unnarrowed t))